requirements.txt
langchain
langchain-openai
langchain_community
langchain-anthropic
langchain-mistralai
langchain-groq
langsmith
langgraph
unstructured[all-docs]
python-docx
streamlit
pdf2docx
pdfminer.six
PyMuPDF
pdfplumber
jinja2
pdfkit
spacy
faker
requests
tavily-python
pandas
